refactor: extract TelemetryEventTypes.swift from TelemetryTypes.swift
Splits TelemetryTypes.swift by concern (mirroring the DiagnosticsTypes
split from #719):
- TelemetryTypes.swift keeps `OperationalTelemetryEvent` (the data
  model that gets recorded)
- TelemetryEventTypes.swift gets `TelemetryEventName` +
  `TelemetryEvent` namespace + the aliases extension

Byte-identical move of lines 15-43. Both files import Foundation.

TelemetryTypes.swift: 43 → 13 lines. Tests: 667.
